requirements.txt
streamlit
pandas
aiohttp
matplotlib
altair
xlsxwriter
Pillow